[Backport] Fix bugs in ModList_RebuildList
Fix bugs in ModList_RebuildList such that it no longer lists
files in the quake directory as gamedirs.

These bugs were previously hidden by a different behavior in FS_CheckNastyPath
until 20140716 where directories with a period in their name were skipped, but
that was not a complete solution.

Thanks to klichka for bug report.

git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@12329 d7cf8633-e32d-0410-b094-e92efae38249