[Backport] Use the invariant keyword on gl_Position
Use the invariant keyword on gl_Position on GLSL 1.20 or higher and
GLSL ES 1.00 or higher, this should resolve zfighting between light
polygons and base surfaces.

A long long time ago this was using ftransform() which ensured invariance
but use of the invariant keyword was not added when we switched away from that.

Thanks for the bug report:
https://www.reddit.com/r/quake/comments/6rwwm5/need_some_help_getting_quake_running_with/

git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@12336 d7cf8633-e32d-0410-b094-e92efae38249

Fix GLSL shader to only specify invariant gl_Position on the vertex shader,
not the fragment shader - woops, this is what I get for not testing
that change.

git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@12337 d7cf8633-e32d-0410-b094-e92efae38249