[Backport]:  Rework R_CanSeeBox a bit to support an eyejitter feature
Rework R_CanSeeBox a bit to support an eyejitter feature, and make it
available throughout the renderer.

Make TraceLineOfSight check if the trace endpos is within a box, not just if fraction is 1 - this makes it much more likely to say an object is visible when the object overlaps walls and other obstacles as the rays only need to reach the box, it doesn't matter much where it goes inside the box.

git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@12348 d7cf8633-e32d-0410-b094-e92efae38249