Implement dynamic ticrate
This feature adds sys_idle_ticrate that saves CPU on empty server.
It's usefull if you going to host large amount of game servers and want
to save some CPU for other applications.

It adds sys_idle_ticrate  cvar, which configures ticrate when server is
empty. When players joins it switched back to sys_ticrate.